% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_uniq.R
\name{len_uniq}
\alias{len_uniq}
\title{len_uniq}
\usage{
len_uniq(vec)
}
\arguments{
\item{vec}{A vector with a various objects}
}
\value{
The number of unique objects in the vector
}
\description{
A short hand for unique() \%>\% length()
}
\examples{
vec <- sample(1:10, size = 20, replace=TRUE)

vec

sort_uniq(vec)
}
