% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{\%!in\%}
\usage{
not_in(x, y)
}
\arguments{
\item{x}{the thing you want to not find in the table}

\item{table}{the table you're searching through for this thing}
}
\value{
boolean
}
\description{
An operator that negates the \%in\% operator
}
\examples{
not_in(1, 2:10)
}
