% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_uniq.R
\name{sort_uniq}
\alias{sort_uniq}
\title{sort_uniq}
\usage{
sort_uniq(vec)
}
\arguments{
\item{vec}{A vector that can be reasonably sorted}
}
\value{
A sorted vector of unique values in the vector
}
\description{
Description
}
\details{
A Shorthand for unique() \%>\% sort()
}
\examples{
vec <- sample(1:10, size = 20, replace=TRUE)

vec

sort_uniq(vec)
}
