% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinasheR-package.R
\docType{package}
\name{tinasheR-package}
\alias{tinasheR}
\alias{tinasheR-package}
\title{tinasheR: My Personal Package}
\description{
A personal package of some go-to code shorthands and frequently used functions I have put together to speed up my development workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TinasheMTapera/tinasheR}
  \item Report bugs at \url{https://github.com/TinasheMTapera/tinasheR/issues}
}

}
\author{
\strong{Maintainer}: Tinashe Tapera \email{tinashemtapera@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
